refactor: テストファイルの移動
- Updated import paths in MLAIService and AIInsights test files for consistency.
- Added trailing commas and adjusted formatting for improved readability and alignment with codebase standards.
